+++ services/linkpreview/unfurlers/unfurler.go
@@ -1,4 +1,4 @@
-package linkpreview
+package unfurlers
 
 import (
 	"context"
