+++ services/linkpreview/unfurlers/unfurler_opengraph.go
@@ -1,4 +1,4 @@
-package linkpreview
+package unfurlers
 
 import (
 	"bytes"
