+++ services/linkpreview/unfurlers/unfurler_status.go
@@ -1,4 +1,4 @@
-package linkpreview
+package unfurlers
 
 import (
 	"fmt"
@@ -10,8 +10,17 @@ import (
 	"github.com/status-im/status-go/images"
 	"github.com/status-im/status-go/protocol/common"
 	"github.com/status-im/status-go/protocol/communities"
+	"github.com/status-im/status-go/protocol/contacts"
 	"github.com/status-im/status-go/services/sharedurls"
 )
+
+//go:generate go tool mockgen -package=mock_unfurlers -source=unfurler_status.go -destination=./mock/unfurler_status.go
+
+type StatusDataProvider interface {
+	GetContactByID(pubKey string) *contacts.Contact
+	FetchContact(contactID string, waitForResponse bool) (*contacts.Contact, error)
+	FetchCommunity(communityID string, shard *messagingtypes.Shard) (*communities.Community, error)
+}
 
 type StatusUnfurler struct {
 	provider StatusDataProvider
